adding CA3

+++ CA2/1.cpp
@@ -4,8 +4,8 @@
 using namespace std;
 
 //this is a functio that give us the number of digit a number
-int num_of_digit(int number){
-    int digits=1;
+long int num_of_digit(long int number){
+    long int digits=1;
     number=abs(number);
     while(number>=10){
         digits++;
@@ -15,7 +15,7 @@ int num_of_digit(int number){
 }
 
 //this function give us a number with all 1 digit
-int make_1_number(int num_of_digit){
+long int make_1_number(long int num_of_digit){
     if(num_of_digit<=1)
         return 1;
     else
@@ -23,16 +23,16 @@ int make_1_number(int num_of_digit){
 }
 
 //this function give us the abs difference of two number
-int dif(int x,int y){
+long int dif(long int x,long int y){
     return abs(abs(x)-abs(y));
 }
 
 //this function find the closest all 1 digits number to a number(closest mean the value of number(abs))
-int find_closest(int number){
-    int n=num_of_digit(number);
-    int num1=make_1_number(n-1);
-    int num2=make_1_number(n);
-    int num3=make_1_number(n+1);
+long int find_closest(long int number){
+    long int n=num_of_digit(number);
+    long int num1=make_1_number(n-1);
+    long int num2=make_1_number(n);
+    long int num3=make_1_number(n+1);
     if(dif(number,num1)<=dif(number,num2) && dif(number,num1)<=dif(number,num3))
         return num1;
     else if(dif(number,num2)<=dif(number,num1) && dif(number,num2)<=dif(number,num3))
@@ -42,7 +42,7 @@ int find_closest(int number){
 }
 
 //this function return the number of digit that we need to make the number(result of problem)
-int min_digit_1_to_made(int input){
+long int min_digit_1_to_made(long int input){
     if(input==0)
         return 0;
     else if(input<0)
@@ -53,7 +53,7 @@ int min_digit_1_to_made(int input){
 }
 
 int main(){
-    int input;
+    long int input;
     cin>>input;
     cout<<min_digit_1_to_made(input);
     return 0;
